requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
pendulum==2.1.2
python-dateutil==2.8.1
pytzdata==2020.1
six==1.15.0
yarl==1.5.1
